Refactor: remove obsolete files and unused features
This commit removes a significant number of obsolete files and unused features to clean up the codebase.

- **Removed Action/Event Filtering UI:**
    - Deleted `ActionCategoriesFiltersActivity`, `ActionLocationFilterActivity`, and their associated layouts and adapters, as this filtering functionality is no longer used in the `ActionsFragment`.
- **Removed Obsolete Activities and Layouts:**
    - Deleted `MyActionsListActivity` and its layout.
    - Deleted `ActivityChooseLanguage` and its layout, as language selection is handled differently.
    - Removed test/temporary activities: `WelcomeTestActivity` and `WebViewActivityForTest` along with their layouts.
- **Refactored Image Viewer:**
    - Renamed and moved image viewing classes from the `discussions` package to a more generic `tools.image_viewer` package.
    - Renamed `ImageDialogActivity` to `ImageViewerActivity` and updated its layout and references for clarity.
- **Asset and Resource Cleanup:**
    - Deleted several unused drawable XML files (`new_profile_header.xml`, `planete.xml`, etc.).
    - Renamed `new_profile_header_orange.xml` to `header_profile_orange.xml` and updated all references to it.
- **Code and String Cleanup:**
    - Removed unused string resources related to the deleted action filtering features across all languages.
    - Removed the "Language" selection item from the "Edit Profile" screen.
    - Added an `EntourageGlideModule` to resolve Glide's `GeneratedAppGlideModule` warning.
