+++ app/src/main/java/social/entourage/android/EntourageGlideModule.kt
@@ -1,0 +1,10 @@
+package social.entourage.android
+
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
+
+@GlideModule
+class EntourageGlideModule : AppGlideModule() {
+    // Leave empty for now. 
+    // This allows Glide to generate the "GeneratedAppGlideModule" it's looking for.
+}
