+++ app/src/main/java/social/entourage/android/home/HomeFragment.kt
@@ -1108,7 +1108,7 @@ class HomeFragment : Fragment(), OnHomeHelpItemClickListener, OnHomeChangeLocati
         var signablePermission = false
         private const val PREF_GATING_ZONE_SHOWN = "PREF_GATING_ZONE_SHOWN"
         private const val PREF_GATING_NOTIF_SHOWN = "PREF_GATING_NOTIF_SHOWN"
-        private const val PREF_GATING_ENHANCED_ONBOARDING_SHOWN =
+        internal const val PREF_GATING_ENHANCED_ONBOARDING_SHOWN =
             "PREF_GATING_ENHANCED_ONBOARDING_SHOWN"
 
         const val PREF_ENHANCED_ONBOARDING_COMPLETED = "PREF_ENHANCED_ONBOARDING_COMPLETED"
