+++ app/src/main/java/social/entourage/android/report/ActionSheetFragment.kt
@@ -16,7 +16,6 @@ import social.entourage.android.api.model.Events
 import social.entourage.android.databinding.NewFragmentSettingsDiscussionModalBinding
 import social.entourage.android.discussions.DetailConversationActivity
 import social.entourage.android.discussions.DiscussionsPresenter
-import social.entourage.android.discussions.imageviewier.ImageListActivity
 import social.entourage.android.discussions.members.MembersConversationFragment
 import social.entourage.android.events.EventsPresenter
 import social.entourage.android.events.create.CreateEventActivity
@@ -29,6 +28,7 @@ import social.entourage.android.profile.ProfileFullActivity
 import social.entourage.android.report.ReportModalFragment
 import social.entourage.android.report.ReportTypes
 import social.entourage.android.small_talks.SmallTalkViewModel
+import social.entourage.android.tools.image_viewer.ImageListActivity
 import social.entourage.android.tools.utils.Const
 import social.entourage.android.tools.utils.CustomAlertDialog
 import social.entourage.android.tools.log.AnalyticsEvents
@@ -387,7 +387,7 @@ class ActionSheetFragment : BottomSheetDialogFragment() {
             if (mode == SheetMode.EVENT && conversationId > 0) {
                 startActivity(
                     Intent(requireContext(), ImageListActivity::class.java)
-                        .putExtra("conversation_id", conversationId)
+                        .putExtra(Const.CONVERSATION_ID, conversationId)
                 )
                 dismiss()
             }
