+++ app/src/main/java/social/entourage/android/tools/TestHelper.kt
@@ -4,6 +4,11 @@ object TestHelper {
     private var isTesting: Boolean? = null
 
     var forceHarness = true
+        set(value) {
+            field = value
+            //force to check class next time we need forceHarness
+            isTesting = null
+        }
 
     fun isRunningInTestHarness(): Boolean {
         if (isTesting == null) {
