--- app/src/main/java/social/entourage/android/welcome/WelcomeTestActivity.kt
@@ -1,49 +1,0 @@
-package social.entourage.android.welcome
-
-import android.content.Intent
-import android.os.Bundle
-import social.entourage.android.base.BaseActivity
-import social.entourage.android.databinding.WelcomeTestLayoutBinding
-
-class WelcomeTestActivity:BaseActivity() {
-
-    private lateinit var binding:WelcomeTestLayoutBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = WelcomeTestLayoutBinding.inflate(layoutInflater)
-        setButtons()
-        setContentView(binding.root)
-
-    }
-
-    private fun setButtons(){
-        binding.welcomeOne.setOnClickListener {
-            val intent = Intent(this, WelcomeOneActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-        binding.welcomeTwo.setOnClickListener {
-            val intent = Intent(this, WelcomeTwoActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-        binding.welcomeThree.setOnClickListener {
-            val intent = Intent(this, WelcomeThreeActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-        binding.welcomeFour.setOnClickListener {
-            val intent = Intent(this, WelcomeFourActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-
-        binding.welcomeFive.setOnClickListener {
-            val intent = Intent(this, WelcomeFiveActivity::class.java)
-            startActivity(intent)
-            finish()
-        }
-    }
-
-}
