--- app/src/main/res/drawable/new_profile_header.xml
@@ -1,38 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="390dp"
-    android:height="844dp"
-    android:viewportWidth="390"
-    android:viewportHeight="844">
-  <path
-      android:pathData="M0,0h390v844h-390z"
-      android:strokeWidth="1"
-      android:fillColor="#FFFCFA"
-      android:fillType="evenOdd"
-      android:strokeColor="#00000000"/>
-  <path
-      android:pathData="M0,0h390v844h-390z"
-      android:strokeWidth="1"
-      android:fillColor="#FFFCFA"
-      android:fillType="evenOdd"
-      android:strokeColor="#00000000"/>
-  <path
-      android:pathData="M0,282.813C52.021,309.842 101.624,344.017 155.554,367.354C228.89,399.088 311.009,409.637 390.161,398.924L390,-0L0.161,-0L0,282.813Z"
-      android:strokeWidth="1"
-      android:fillColor="#FFEADC"
-      android:fillType="evenOdd"
-      android:strokeColor="#00000000"/>
-  <path
-      android:pathData="M263.935,44.764C261.756,29.436 271.369,-0.274 271.369,-0.274C271.369,-0.274 310.912,-0.183 390,0C390.108,47.39 390.161,73.532 390.161,78.425C390.161,85.765 390.198,92.927 390,93.063C381.874,98.678 376.065,100.35 369.014,103.298C360.073,107.038 350.787,109.971 341.183,111.271C337.96,111.709 332.44,111.821 323.773,110.095C293.21,102.972 268.354,75.833 263.935,44.764Z"
-      android:strokeAlpha="0.25"
-      android:strokeWidth="1"
-      android:fillColor="@color/orange"
-      android:fillType="evenOdd"
-      android:strokeColor="#00000000"
-      android:fillAlpha="0.25"/>
-  <path
-      android:pathData="M0,248.267C1.317,248.914 2.303,249.394 2.959,249.704C65.731,279.484 114.961,257.119 164.31,232.936C184.366,223.104 202.43,210.17 218.296,194.9C269.374,145.738 296.308,71.077 296.309,0.3C296.309,0.1 197.592,0 0.161,0L0,248.267Z"
-      android:strokeWidth="1"
-      android:fillColor="#FF9C5D"
-      android:fillType="evenOdd"
-      android:strokeColor="#00000000"/>
-</vector>
