--- app/src/main/res/drawable/planete.xml
@@ -1,13 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="20dp"
-    android:height="20dp"
-    android:viewportWidth="20"
-    android:viewportHeight="20">
-  <group>
-    <clip-path
-        android:pathData="M0,0h20v20h-20z"/>
-    <path
-        android:pathData="M2.254,7.539L2.621,8.18C2.945,8.746 3.477,9.164 4.105,9.344L6.367,9.988C7.039,10.18 7.5,10.793 7.5,11.492V13.051C7.5,13.481 7.742,13.871 8.125,14.063C8.508,14.254 8.75,14.644 8.75,15.074V16.598C8.75,17.207 9.332,17.648 9.918,17.48C10.547,17.301 11.035,16.797 11.195,16.16L11.305,15.723C11.469,15.063 11.898,14.496 12.488,14.16L12.805,13.981C13.391,13.648 13.75,13.023 13.75,12.352V12.027C13.75,11.531 13.551,11.055 13.199,10.703L13.047,10.551C12.695,10.199 12.219,10 11.723,10H10.039C9.605,10 9.176,9.887 8.797,9.672L7.449,8.902C7.281,8.805 7.152,8.648 7.09,8.465C6.965,8.09 7.133,7.684 7.488,7.508L7.719,7.391C7.977,7.262 8.277,7.238 8.551,7.332L9.457,7.633C9.777,7.738 10.129,7.617 10.313,7.34C10.496,7.066 10.477,6.703 10.266,6.449L9.734,5.813C9.344,5.344 9.348,4.66 9.746,4.199L10.359,3.484C10.703,3.082 10.758,2.508 10.496,2.051L10.402,1.887C10.266,1.879 10.133,1.875 9.996,1.875C6.371,1.875 3.297,4.254 2.254,7.539ZM18.125,10C18.125,8.563 17.75,7.211 17.094,6.035L16.094,6.438C15.481,6.684 15.164,7.367 15.371,7.992L16.031,9.973C16.168,10.379 16.5,10.688 16.914,10.789L18.051,11.074C18.098,10.723 18.121,10.363 18.121,10H18.125ZM0,10C0,7.348 1.054,4.804 2.929,2.929C4.804,1.054 7.348,0 10,0C12.652,0 15.196,1.054 17.071,2.929C18.946,4.804 20,7.348 20,10C20,12.652 18.946,15.196 17.071,17.071C15.196,18.946 12.652,20 10,20C7.348,20 4.804,18.946 2.929,17.071C1.054,15.196 0,12.652 0,10Z"
-        android:fillColor="#FF9739"/>
-  </group>
-</vector>
