--- app/src/main/res/layout/activity_action_cat_filters.xml
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/action_cat_filters_view"
-    tools:context=".actions.ActionCategoriesFiltersActivity">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@color/light_beige">
-
-        <LinearLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <include
-                android:id="@+id/header"
-                layout="@layout/layout_header"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                app:layout_constraintBottom_toTopOf="@id/recycler_view"
-                app:layout_constraintTop_toTopOf="parent"
-                app:title="@{@string/action_cat_filters_title}" />
-
-            <androidx.recyclerview.widget.RecyclerView
-                android:id="@+id/recycler_view"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="16dp"
-                tools:itemCount="5"
-                tools:listitem="@layout/layout_action_category_item" />
-        </LinearLayout>
-
-        <include
-            android:id="@+id/validate"
-            layout="@layout/new_rounded_button_white"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginHorizontal="65dp"
-            android:layout_marginBottom="32dp"
-            app:content="@{@string/validate}"
-            app:drawable="@{@drawable/new_rounded_button_orange}"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
