--- app/src/main/res/layout/activity_choose_language_layout.xml
@@ -1,101 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <TextView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_fr"
-        android:fontFamily="@font/nunitosans_bold"
-        android:textColor="@color/orange"
-        android:textSize="16sp"
-        android:layout_margin="16dp"
-        android:text="\uD83C\uDDEB\uD83C\uDDF7 Français"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
-
-    <TextView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_en"
-        android:text="English"
-        android:layout_margin="16dp"
-        android:fontFamily="@font/nunitosans_regular"
-        android:textColor="@color/orange"
-        android:textSize="16sp"
-        app:layout_constraintTop_toBottomOf="@+id/tv_fr"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-    <TextView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_de"
-        android:text="Deutsch"
-        android:layout_margin="16dp"
-        android:fontFamily="@font/nunitosans_regular"
-        android:textColor="@color/orange"
-        android:textSize="16sp"
-        app:layout_constraintTop_toBottomOf="@+id/tv_en"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-    <TextView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_pl"
-        android:text="Polski"
-        android:layout_margin="16dp"
-        android:fontFamily="@font/nunitosans_regular"
-        android:textColor="@color/orange"
-        android:textSize="16sp"
-        app:layout_constraintTop_toBottomOf="@+id/tv_de"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-    <TextView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_ro"
-        android:text="Română"
-        android:layout_margin="16dp"
-        android:fontFamily="@font/nunitosans_regular"
-        android:textColor="@color/orange"
-        android:textSize="16sp"
-        app:layout_constraintTop_toBottomOf="@+id/tv_pl"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-    <TextView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_uk"
-        android:text="Українська"
-        android:layout_margin="16dp"
-        android:fontFamily="@font/nunitosans_regular"
-        android:textColor="@color/orange"
-        android:textSize="16sp"
-        app:layout_constraintTop_toBottomOf="@+id/tv_ro"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-    <TextView
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:id="@+id/tv_ar"
-        android:text="العربية"
-        android:layout_margin="16dp"
-        android:fontFamily="@font/nunitosans_regular"
-        android:textColor="@color/orange"
-        android:textSize="16sp"
-        app:layout_constraintTop_toBottomOf="@+id/tv_uk"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:textAlignment="textEnd"
-        />
-
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
