+++ app/src/main/res/layout/create_survey_activity.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:background="@drawable/new_profile_header_orange"
+    android:background="@drawable/header_profile_orange"
     android:fitsSystemWindows="true">
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/header"
