+++ app/src/main/res/layout/fragment_association_profile.xml
@@ -7,7 +7,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@drawable/new_profile_header">
+        android:background="@drawable/header_profile_orange">
 
         <ImageView
             android:id="@+id/asso_profile_image_association"
