+++ app/src/main/res/layout/fragment_create_action.xml
@@ -8,7 +8,7 @@
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@drawable/new_profile_header_orange"
+        android:background="@drawable/header_profile_orange"
         android:orientation="vertical">
 
         <androidx.constraintlayout.widget.ConstraintLayout
