+++ app/src/main/res/layout/fragment_create_event.xml
@@ -9,7 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:id="@+id/create_event_layout"
-        android:background="@drawable/new_profile_header_orange"
+        android:background="@drawable/header_profile_orange"
         android:orientation="vertical" >
 
         <include
