+++ app/src/main/res/layout/fragment_groups.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:background="@drawable/new_profile_header_orange_dark"
+    android:background="@drawable/header_profile_orange"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".groups.GroupsFragment">
