+++ app/src/main/res/layout/fragment_image_viewer.xml
