+++ app/src/main/res/layout/fragment_messages.xml
@@ -7,7 +7,7 @@
     <androidx.coordinatorlayout.widget.CoordinatorLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@drawable/new_profile_header_orange">
+        android:background="@drawable/header_profile_orange">
 
         <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/appBar"
