+++ app/src/main/res/layout/new_fragment_edit_group.xml
@@ -7,7 +7,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:background="@drawable/new_profile_header_orange"
+        android:background="@drawable/header_profile_orange"
         android:orientation="vertical">
 
         <include
