+++ app/src/main/res/values-ar/strings.xml
@@ -593,22 +593,14 @@
     <string name="actions_tab_contribs">مساهمات</string>
     <string name="actions_tab_demands">طلب مساعدة</string>
 
-    <string name="action_bt_my_actions">إعلاناتي</string>
-
     <string name="action_contrib_empty_state_title">لا توجد مساهمات تطابق بحثك</string>
     <string name="action_contrib_empty_state_desc">يمكنك توسيع مجال البحث الخاص بك أو إنشاء مساهمة</string>
 
     <string name="action_demand_empty_state_title">لا توجد طلبات تطابق بحثك</string>
     <string name="action_demand_empty_state_desc">يمكنك توسيع حقل البحث الخاص بك أو إنشاء طلب</string>
 
-    <string name="action_bt_cat_filters">فئات</string>
-    <string name="action_cat_filters_title">تصفية حسب الفئات</string>
-    <string name="action_location_filter_title">التصفية حسب الموقع</string>
-
     <string name="action_menu_create_demand">إنشاء طلب</string>
     <string name="action_menu_create_contrib">إنشاء مساهمة</string>
-
-    <string name="action_my_actions">إعلاناتي</string>
 
     <string name="action_my_empty_title">لم تقم بنشر أي إعلانات</string>
     <string name="action_my_empty_subtitle">يمكنك إنشاء مساهمة أو طلب بالضغط على "+" في صفحة "المساعدة المتبادلة".</string>
