+++ app/src/main/res/values-en/strings.xml
@@ -542,17 +542,13 @@
     <string name="actions_tab_demands">Requests</string>
     <string name="action_date_list">%s</string>
 
-    <string name="action_bt_my_actions">My listings</string>
     <string name="action_contrib_empty_state_title">No contributions match your search</string>
     <string name="action_contrib_empty_state_desc">You can broaden your search field or create a contribution</string>
     <string name="action_demand_empty_state_title">No requests match your search</string>
     <string name="action_demand_empty_state_desc">You can expand your search field or create a request</string>
-    <string name="action_bt_cat_filters">Categories</string>
-    <string name="action_cat_filters_title">Filter by categories</string>
-    <string name="action_location_filter_title">Filter by location</string>
+
     <string name="action_menu_create_demand">Create request</string>
     <string name="action_menu_create_contrib">Create contribution</string>
-    <string name="action_my_actions">My listings</string>
     <string name="action_my_empty_title">You have not published any listings</string>
     <string name="action_my_empty_subtitle">You can create a contribution or request by clicking on the “+” on the “Mutual aid” page.</string>
     <string name="action_cancel_title">The listing has been deleted...</string>
