+++ app/src/main/res/values/strings.xml
@@ -628,22 +628,14 @@ Ils liront ce message dès qu’ils rejoindront le groupe.</string>
 
     <string name="action_date_list">%s</string>
 
-    <string name="action_bt_my_actions">Mes annonces</string>
-
     <string name="action_contrib_empty_state_title">Aucune contribution ne correspond à votre recherche</string>
     <string name="action_contrib_empty_state_desc">Vous pouvez élargir votre champ de recherche ou créer une contribution</string>
 
     <string name="action_demand_empty_state_title">Aucune demande ne correspond à votre recherche</string>
     <string name="action_demand_empty_state_desc">Vous pouvez élargir votre champ de recherche ou créer une demande</string>
 
-    <string name="action_bt_cat_filters">Catégories</string>
-    <string name="action_cat_filters_title">Filtrer par catégories</string>
-    <string name="action_location_filter_title">Filtrer par localisation</string>
-
     <string name="action_menu_create_demand">Créer une demande</string>
     <string name="action_menu_create_contrib">Créer une contribution</string>
-
-    <string name="action_my_actions">Mes annonces</string>
 
     <string name="action_my_empty_title">Vous n’avez publié aucune annonce</string>
     <string name="action_my_empty_subtitle">Vous pouvez créer une contribution ou une demande en cliquant sur le « + » de la page « Entraide »</string>
